Vignette, Camera Shake and Lock on Baby
Many changes but they all look small
The camera can shake in a for loop then also does nested for loop for turn off
Lock on now has an indicator based on the enemy. Darksouls style so it doesn't get bigger or small, can change if you want me to.
Adjustable vignette with exposed strength and beginning variable.